tingleProject/accounts/views.py
from .models import CustomUser
from .serializers import UserSerializer, MyTokenObtainPairSerializer
from rest_framework import viewsets
from rest_framework_simplejwt.views import TokenObtainPairView

class UserViewSet(viewsets.ModelViewSet):
    queryset = CustomUser.objects.all()
    serializer_class = UserSerializer

class MyTokenObtainPairView(TokenObtainPairView):
    serializer_class = MyTokenObtainPairSerializer